+++ examples/typescript-esm/simple_digital_asset.ts
@@ -5,7 +5,8 @@
  * This example shows how to use the Aptos client to mint and transfer a Digital Asset.
  */
 
-import "dotenv";
+import dotenv from "dotenv";
+dotenv.config();
 import { Account, Aptos, AptosConfig, Network, NetworkToNetworkName } from "@aptos-labs/ts-sdk";
 
 const INITIAL_BALANCE = 100_000_000;