+++ examples/typescript/public_key_authenticator_account_abstraction.ts
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
-
+import dotenv from "dotenv";
+dotenv.config();
 import {
   Account,
   AbstractedAccount,